Now can run the tests on a mysql ram disk server.
As part of trying to get it so that tests can be run very easily, I've added
a mode to bin/civiprj that will create the project and run the tests on it.

Also added an option to have everything run on top of a MySQL server that has
stores everything on a RAM disk which greatly speeds up the tests.