civilint - Allow execution in a git subdir
If you check out a git repo, go to a subdir, edit a file, and run
`civilint`, then you would expect it to scan that file...  but it actually
fails.

There is a subtlety in what the behavior should be -- e.g.  does it work
like `git status` or `git diff` (which check the whole repo) or like `find`
or `grep -r` (which look under the current folder).

As it happens, it's much easier to implement the first behavior, so that's
what this PR does...