PCHR-1633: Created Patch for Mailsystem 7.x-2.4
Latest version of MailSystem module was breaking the application on creating a new Mail class using
admin interface. The problem was, when either private:// or public:// resolved to a route outside
of drupal root directory, the absolute path of the file was being saved to registry, breaking the
application on class instantiation.  Solved by calculating route relative to drupal root path in
a patch that now has to be applied on CiviHR installation.

Patch for latest development branch of MailSystem module failed to apply to version being used
on CiviHR, which is 7.x-2.34.  Created patch for this version and fixed array of patches for
installation.