composer.json - Remove extra copy of phpunit (et al)
Background
----------

CiviCRM's usage of PHPUnit evolved along this track:

1. In `civicrm-core` and `civicrm-packages`, there was a forked copy of PHPUnit 3.x. It had various patches applied,
   and the test suite didn't run with vanilla PHPUnit. (Things ran this way for several years.)
2. Added vanilla phpunit 3.x to `civicrm-buildkit` (via `composer`). Patched `civicrm-core` until it's compatible.
   (Things didn't run this way as long -- maybe a few months?)
3. Added vanilla phpunit 4.x to `civicrm-buildkit` (via `phar`). Patched `civicrm-core` until it's compatible.
   (Things have now run this way a couple years.)

The take-away is that the composer-based copy of phpunit 3.x hasn't been used
for a long time.  (Not by our CI, anyway.) There's been some confusion about
this (e.g.  because the old version has the default command name `phpunit`).

Following the [discussions on #390 and #345](https://github.com/civicrm/civicrm-buildkit/pull/390#discussion_r186237206),
I'm submitting a replacement PR which removes the unused PHPUnit 3.x.

Before
------

* `bin/phpunit4` and `bin/phpunit5` are the main binaries used by CI and docs
* `bin/phpunit` is a symlink to `vendor/phpunit/phpunit`, and `vendor/phpunit/*` contains an ancient build

After
-----

* `bin/phpunit4` and `bin/phpunit5` are the main binaries used by CI and docs
* `bin/phpunit` and `vendor/phpunit/*` do not exist

Comments
--------

After removing this copy of phpunit on my local, I successfully ran
the test suites for 4.6 and 5.x (to confirm the expectation that it was
unused).

The other packages removed in this PR are phpunit-derivatives.  dbunit and
selenium were included to get extra helper classes into the phpunit 3.x
runtime.  paratest was an experimental effort to do multiprocess testing.