civi-download-tools - Rename PHPUnit files. Improve PHPStorm compatibility.
Overview
--------

When using PHPStorm, you need to setup a few things, including:

* Registering the phpunit binary (for execution)
* Registering phpunit as part of the "include path" (for class-loading/identification)

However, this wasn't working with `bin/phpunit4`-- because PHPStorm expects that the
binary has the literal extension `.phar`, and it can't scan effectively if
multiple versions of phpunit are in the same folder.

This patch renames the files and leaves symlinks.

Before
------

* `bin/phpunit4` is a downloaded PHAR file
* `bin/phpunit5` is a downloaded PHAR file

After
-----

* `bin/phpunit4` is a symlink to `extern/phpunit4/phpunit4.phar`
* `bin/phpunit5` is a symlink to `extern/phpunit5/phpunit5.phar`
* `extern/phpunit4/phpunit4.phar` is a downloaded PHAR file
* `extern/phpunit5/phpunit5.phar` is a downloaded PHAR file