drush - Fix random PHAR exceptions. Use dedicated cache dir.
After deploying the drush8 PHAR, we started getting random failures during builds.
This appears to happen because different copies of `drush8.phar`
(eg `bknix-min/bin/drush8` and `bknix-max/bin/drush8`) were writing data to the
same cache folder (`~/.drush`) ... but with conflicting metadata.

The specific symptom involved an error from the PharInterceptor, which I believe worked
like this:

* `bknix-min/bin/drush8` would write metadata about available commandfiles to `~/.drush`
* `bknix-max/bin/drush8` would read the metadata from`~/.drush`; but tha would point to
  commandfiles nested under `bknix-min/bin/drush8`
* `bknix-max/bin/drush8` would try to load the commandfiles from `bknix-min/bin/drush8`
* This would require opening the PHAR, but the PHAR had a non-PHAR extension.
* This would provoke an exception from D7's customized `phar://` stream wrapper.

This was very hard to diagnose because the symptoms did not appear
consistently -- if you had a dozen test jobs using a mix of environments,
then the happenstance of job-ordering and cache-timeouts would mean that
you'd only provoke a problematic crossover half the time.

This patch forces buildkit's copy of `drush` to use a separate cache folder
under `app/tmp/drush`.