civibuild upgrade-test - Fix support for "web/"
The variables "WEB_ROOT" and "CMS_ROOT" are named confusingly.
In the past, they were the same on D7 sites, so it didn't matter,
but now it does.

(Distinction: WEB_ROOT is actually the root of the specific build;
CMS_ROOT is the http document root served by the cms)