cvutil_inject_settings() - Soften reliance on CMS_VERSION
The CMS_VERSION was added to the list of required inputs so that it could be
passed through to `$civibuild['CMS_VERSION']` and then eventually to
`app/drupal.settings.d/pre.d/100-file_private_path-d8.php`.

Evidentally (from #503) some old builds did not record a `CMS_VERSION`, so
the assertion raises a fatal.  I'm not certain why - but regardless, the
situation should not be a fatal for most build types (since the value
normally isn't even used).