bin/civi-git-table - Tweak order
I use `civi-git-table` to ensure that the candidate tarballs are up-to-date
with the corresponding git branches (i.e.  each repo has the same revision
that was used in the tarball).

It's easier if the information is presented in a consistent order.

Compare with https://download.civicrm.org/latest/branch/X.X/LATEST/civicrm-X.Y.Z-drupal-LATEST.tar.gz/inspect