loco - For CI configuration profiles, use 'environment:' instead of 'default_environment:'
The loco.yml format offers a distinction between 'environment:' and
'default_environment:' under the expectation that a user may want
to intervene and adjusting certain settings at runtime:

```
env HTTPD_DOMAIN=customized loco run
```

This makes sense if the user is manually running the command.

However, in the CI deployments, we use the YAML to generate systemd-units
and run everything in the background.  There is no user who launching the command
and no environment through which they can intervene.

At a high level, using 'default_environment:' for CI configs just muddies
the water - it adds a question mark to what the value will actually be. There's
no need to improvise the settings.

Additionally, there's a functional issue. Say I'm on a developer workstation and run this:

```
me@localhost$ use-bknix min -s
[bknix-min:~]$ loco run -f .loco/jenkins-dfl.yml
```

As a user, I expect it to follow the options in `.loco/jenkins-dfl.yml`. But if
`jenkins-dfl.yml` is deferrential, then we won't actually use any the constants from there.
Why? Because `use-bknix` initialized the environment with the default config (`loco.yml`),
and now that environment will supercede any defaults.

With this change, there's less room for error when using these profiles.