When using nix-shell, the loco variables should be exported
Suppose you open a `nix-shell -A min`.  The env-var `HTTPD_PORT` appears to be defined.
However, if you open a subshell (`bash`), you'll discover that `HTTPD_PORT` has
not propagated to the next command.

On the other hand, if you run `use-bknix min -s` with a similar flow, the `HTTPD_PORT` is
propagated.

This patch brings `nix-shell` into closer alignment with `use-bknix`.