infra/ops#955 - Add bknix-ci-cleanup
This script aims to degrade more gracefully when the hardware capacity is
inadequate to providing the nominal 5 days of support for a test build.

This script will start by deleting 5-day-old builds.  If the disk still
looks too full, then it proceeds to 4-day-old builds.  Then 3-day (etc).

(The definition of "too full" is "using more than 90% of disk-space or inodes".)