bknix-ci-cleanup - Be quiet about /tmp oddballs
This is a bit unfortunate, but I don't see a simple alternative.

The intent of `find /tmp -user jenkins -delete` is to only match files owned by
`jenkins`.  The problem is that you might have a folder like `/tmp/some-root-user-cache`.
While scanning for files owned by `jenkins`, it will attempt to scan
`/tmp/some-root-user-cache`, but that causes a permission failure.  The `find` scan will
continue anyway - but exits with a failure at the end.

Ideally, we'd ignore `/tmp/some-root-user-cache` completely and still allow other errors
(disk IO errors) to raise an alarm.  Alas, I don't see how to do that with `find`.  So
we're just going to subdue all errors from `find` -- because:

- The `/tmp/some-root-user-cache`-type scenario is quite common.
- The disk IO-type scenario is quite uncommon.