nix/pkgs/php71+ - Enable APC(u) backward-compatibility extension
APC(u) provides an API for in-memory data-caching.  The extension is a
pre-requisite for running `E2E_Cache_APCcacheTest`.

Small clarification about "APC(u)" naming:

* `apc` was the original extension. It provided both data-caching and opcode-caching.
* `apcu` is a successor in PHP 5.6/7.x which provides only data-caching.
* `apcu_bc` is a backward-compatibility extension. It supports the `apc_*` APIs on
   top of the `apcu` extension.

By default, APC(u) is only enabled for web-requests. The php.net docs suggest it's
a waste of resources for most CLI uses. However, it *can be* for testing. With this
revision, APC(u) is generally disabled for CLI work, but it may be enabled via
env-var PHP_APC_CLI.