civi-test-run - Enable APC(u) for E2E test-suite
APC(u) is disabled by default in most PHP-CLI variants.  This means that
`E2E_Cache_APCcacheTest` is often skipped.

In the bknix's PHP, you can use an env-var to enable it. That's all we need.

(This wouldn't apply to other environments...  but they're no worse off than
before...  because the env-var would be ignored elsewhere.  In any case,
it's one small/low-visibility test, and there's not a lot we can do to setup
APCu on other environments.)