php73 - Fix PHAR loading issue. Update php73 to 7.3.33.
The existing build of php73 does not run some PHARs (notably, the current
`composer.phar`).  Why?  PHARs can be signed with different hashing
algorithms, and this one was signed with SHA-512.  Apparently, the PHAR
signature-checking rejects SHA-512 if the `hash` library is dynamically
linked.

With this PR, the `hash` will be built into PHP (which fixes PHAR
signature-checking).  For good measure, it also updates to PHP 7.3.33.