civi-test-run - Add extra flushes between test-runs
Context: We recently made `search_kit` mandatory and wanted to update the `drupal-clean` profile
to use the standard installer (which will enable `search_kit` sooner).

However, `civi-test-run phpunit-core-exts` exhibits a weird failure in this
mode (see #735).  The failure can be reproduced most quickly by comparing:

  cd ext
  (civibuild restore && cd civiimport && phpunit8 --group headless )
  (civibuild restore && cd search_kit && phpunit8 --group headless )

But if you throw in an extra `cv flush` after resetting the DB, it works fine. This
is strange, and it doesn't seem to be as simple as `templates_c`-clash. Still
not sure what resource is conflicted.

Regardless, this script is only used in CI, so the extra second of runtime
won't be noticed.