Get packages from the "default" pin
This is probably the most unusual part of civicrm-buildkit's nix. Compare:

(1) Typical first line:

   { goPackages }:

This declares the overall file as a *function*.  `goPackages` is a parameter
that we expect.  It's somebody else's responsibility to find it. There are
good reasons why this is normal, but it relies on other stuff that's a bit
magical for me.

(2) In civicrm-buildkit, you'll often see things like:

   with (import ../../pins).v2205;

This loads a specific release of nixpkgs and makes all of it available.

There are good reasons why this is less normal, but it's a lot easier for
me to test/use.

These two examples reflect circumstance: many authors are submitting
packages for distribution through "nixpkgs".  Here, we're positioned
downstream - juggling multiple versions of "nixpkgs" (so that we can
mix/match with new+old versions of various packages).