civi-test-run - Don't override TMPDIR
This line was added by cf883ed61236aa7711e6f679ef0389c6578f5e29; i.e.  it
was general/global change that added similar statements in several files.
However, if you look in this file specifically, it never uses TMPDIR.

This line manifested as a problem while setting up `test-4`. If you
set `TMPDIR` to `buildkit/build/.civibuild/tmp`, then the path has
a dot. Some tests in `phpunit-crm` create temp data and then perform
scans on it (e.g.`findFiles()`). That fails due to the dot in `.civibuild`.

We should probably fix `findFiles()`, but for the interim `test-4` needs to
work with existing branches/releases. Even if that's fixed today, it would
take a few months for affected versions to age-out.