nix - Declare `coreutils` requirement
coreutils is the GNU package providing commands like `ls`,`chmod`, `df`, and
so on.

It's generally implicit that all the environments running buildkit have
these commands.  However, at one point I found it handy to specifically load
this.  (I think I was playing around with Docker builds of the bknix
profiles?  Or maybe using `--pure`?) So I've had this patch sitting on my
laptop for a while.

In any event, the effect of this patch should be that bknix environments will use
specific versions of `ls`, `chmod`, etc as per nixpkgs (rather than
inheriting the versions from the host).