standalone-composer - Fix support for `--civi-ver` flag
Overview
--------

I was recently surprised to see new test failures on the 5.75-stable test-matrix
regarding Standalone... even though there were no relevant changes. The problem?
It wasn't actually running tests on 5.75. It was running them on master (regardless
of requested version).

This updates `standalone-composer` to respect the requested version.

Steps to Reproduce
------------------

```
civibuild create standalone-composer --civi-ver 5.75`
```

Before
------

It ignores the requested version (5.75) and installs `master`.

After
-----

IT respects the requested version. (But in absence of a version, it preserves the default `master`.)