(infra/ops#1025) releaser.php - Swap target URL
When submitting an upload to Gitlab, it responds with info about the new upload:

{
    "alt": "civicrm-5.75.2-backdrop.tar.gz",
    "url": "/uploads/XXXXXXXX/civicrm-5.75.2-backdrop.tar.gz",
    "full_path": "/-/project/XXX/uploads/XXXXXXXX/civicrm-5.75.2-backdrop.tar.gz",
    "markdown": "[civicrm-5.75.2-backdrop.tar.gz](/uploads/XXXXXXXX/civicrm-5.75.2-backdrop.tar.gz)"
}

On new uploads, it still returns a `url`, but (for whatever reason) the
`url` doesn't actually work like before.

However, the `full_path` is working on new uploads. So we'll use that...

(This requires constructing the URL relative to a different base -- eg `url`
was project-relative, and `full_path` is domain-relative.)