drupal{9,10}-dev - Enable symlink for core code
BEFORE:

* Create ./src/civicrm-core and apply patches
* When installing dependencies, `composer` identifies `./src/civicrm-core`
* And then it copies the files

AFTER:

* Create ./src/civicrm-core and apply patches
* When installing dependencies, `composer` identifies `./src/civicrm-core`
* And then it makes one big symlink

This is a generally appealing behavior. However, the immediate issue is that
(if we apply core PR 31498) then the "copy" mode will no longer the test-suite.
And so we can't run tests.

The symlink ensures we have access to the test files.

COMMENT

Using the symlink mode requires fixes in civicrm-core, so it's only used for 5.81+.