bknix-ci-cleanup - Quick fix for PHP-warning-cascade
1. To run the cleanup script, you need an environment with PHP. (Call this PHP[A].)
2. Within the cleanup script, it switches between multiple environments with several different PHPs. (Call these PHP[B,n])
3. When switching from PHP[A] to PHP[B,n], some artifact leaks between them.
4. This has traditionally caused PHP_B[n] to print warnings.
5. PHP warnings sometimes cascade into fatal problems.

The mechanics of loading PHP[n] vary by package-manager/version, but nix-php
often uses some mix of env-vars to give different hints to each PHP[n]. But
env-vars are sticky. Vars intended for PHP[A] wind up leaking to PHP[B,n].

The approach in this patch is to attack the hand-off moment, where PHP[A]
calls out to PHP[B,n]. We reset the entire environment to a pristine baseline
(without any preference for any PHP).

Usage:

  ## Start session
  export -p > $HOME/my-baseline-env
  USE_PHP_A...

  ## Run cleanup script
  bknix-ci-cleanup run --env-file="$HOME/my-baseline-env" ...

This ensures that all interpreters start with the same pristine environment.

This is a short-term measure for use with the `install-ci.sh`
system-configuration. In newer configurations (like `install-runner.sh` and
`install-demo.sh`), this will be moot. (They have better switching mechanism -- eg
`homerdo`, `ssh`, `systemd-run`. And they have different lifecycles and quotas.)