phpunit-xml-cleanup - Re-import code
This was originally a helper script within `civicrm-buildkit.git`. It needed
to grow to include some tests+examples, and (at the time) there was no
good place to store those in `civicrm-buildkit.git`. So it moved off to another repo.

Following #948, there is now a better place under `civicrm-buildkit.git:tools/`.

Keeping `phpunit-xml-cleanup` in a separate repo would make it a bit of an outlier.
It's more consistent to bring into `tools/`