civi-download-tools - Optimize conditionals
In bash, the expressions `[...]` and `[[...]]` are similar-but-different.

The single `[` spawns a subprocess, but the double `[[` is evaluated in the
same process. So double `[[` is faster.

For simple expressions, they work the same way.  But as the boolean
expressions get more complex, they start to differ.

So I've changed all the simple ones.