CiviCRM-Ext-Matrix - Skip old 'npm' command
It appears that the purpose of this line was to force `npm` to store its temporary
data somewhere under HOME (rather than /tmp).

We recently updated to a newer versions node and npm for CI -- and this line started failing.

It seems that line may be obsolete -- npm now stores more data under HOME by default...

https://medium.com/@ruben.alapont/npm-caching-speeding-up-your-development-process-340dcdc554b3

So I guess it's worth a try to run without these older lines.