nix - Fix for installing profile snapshots (re:use-bknix) on new hosts
This bit of code was written to bridge a breaking-change in a past version
of nix. Basically, they moved one of the folders that's important for
`use-bknix` -- and so you have to detect what the proper path is.

The detection works by checking your existing folders -- and that's fine if
you already have some profiles (either from past bknix or from something
else). However, on a really new system, there are no folders -- and
it fails to pick the right base.

With this update, `install-developer.sh` should work for:

* Existing installations (with old layout)
* Existing installations (with new layout)
* New installations (with new layout)